+++ openapi_python_client/__init__.py
@@ -96,7 +96,8 @@ class Project:
 
         self.package_name: str = self.package_name_override or self.project_name.replace("-", "_")
         self.package_dir: Path = self.project_dir / self.package_name
-        self.package_description = f"A client library for accessing {self.openapi.title}"
+        self.package_description: str = f"A client library for accessing {self.openapi.title}"
+        self.version: str = openapi.version
 
         self.env.filters.update(self.TEMPLATE_FILTERS)
 
@@ -158,7 +159,10 @@ class Project:
         pyproject_path = self.project_dir / "pyproject.toml"
         pyproject_path.write_text(
             pyproject_template.render(
-                project_name=self.project_name, package_name=self.package_name, description=self.package_description
+                project_name=self.project_name,
+                package_name=self.package_name,
+                version=self.version,
+                description=self.package_description,
             )
         )